:arrow_down: Using Javax instead of Jakarta (Hibernate).

+++ src/main/java/com/bsfdv/transaction/model/Role.java
@@ -1,6 +1,6 @@
 package com.bsfdv.transaction.model;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -14,7 +14,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
